feat: saving data in memory and separating logic
