--- api/models/hello_controller.py
@@ -1,4 +1,0 @@
-
-
-def hello_world():
-    return 'Hello, World!'
