+++ example.http
@@ -2,7 +2,7 @@ POST http://localhost:3000/api/bids HTTP/1.1
 Content-Type: application/json
 
 {
-  "id": "test_id",
+  
   "tender": "test_tender",
   "client": "test_client",
   "alias": "test_alias",
@@ -10,8 +10,8 @@ Content-Type: application/json
   "bid_folder_url": "test_folder_url",
   "status": "in-progress",
   "links": {
-    "self": "/bids/test_id",
-    "questions": "/bids/test_id/questions"
+    "self": "/bids/",
+    "questions": "/bids//questions"
   },
   "was_successful": "test_success",
   "success": [
