+++ test.txt
@@ -1,0 +1,1 @@
+{'id': UUID('60309da9-73a1-4e9c-a647-09d6310be535'), 'tender': 'test_tender', 'client': 'test_client', 'alias': 'test_alias', 'bid_date': '2023-06-21T10:00:00Z', 'bid_folder_url': 'test_folder_url', 'status': 'in-progress', 'links': {'self': '/bids/60309da9-73a1-4e9c-a647-09d6310be535', 'questions': '/bids/60309da9-73a1-4e9c-a647-09d6310be535/questions'}, 'was_successful': 'test_success', 'success': [{'phase': 1, 'has_score': True, 'score': 80, 'out_of': 100}], 'failed': {'phase': 2, 'has_score': False}, 'feedback': {'description': 'test_description', 'url': 'test_url'}, 'last_updated': ''}{'id': UUID('6e8607cb-12af-4dde-8ed2-b15786a5c1f8'), 'tender': 'test_tender', 'client': 'test_client', 'alias': 'test_alias', 'bid_date': '2023-06-21T10:00:00Z', 'bid_folder_url': 'test_folder_url', 'status': 'in-progress', 'links': {'self': '/bids/6e8607cb-12af-4dde-8ed2-b15786a5c1f8', 'questions': '/bids/6e8607cb-12af-4dde-8ed2-b15786a5c1f8/questions'}, 'was_successful': 'test_success', 'success': [{'phase': 1, 'has_score': True, 'score': 80, 'out_of': 100}], 'failed': {'phase': 2, 'has_score': False}, 'feedback': {'description': 'test_description', 'url': 'test_url'}, 'last_updated': ''}{'id': UUID('c465d845-f677-4700-8236-fad15dd12859'), 'tender': 'test_tender', 'client': 'test_client', 'alias': 'test_alias', 'bid_date': '2023-06-21T10:00:00Z', 'bid_folder_url': 'test_folder_url', 'status': 'in-progress', 'links': {'self': '/bids/c465d845-f677-4700-8236-fad15dd12859', 'questions': '/bids/c465d845-f677-4700-8236-fad15dd12859/questions'}, 'was_successful': 'test_success', 'success': [{'phase': 1, 'has_score': True, 'score': 80, 'out_of': 100}], 'failed': {'phase': 2, 'has_score': False}, 'feedback': {'description': 'test_description', 'url': 'test_url'}, 'last_updated': ''}{'id': UUID('99f38b86-ae2c-405c-8508-a897a3521551'), 'tender': 'test_tender', 'client': 'test_client', 'alias': 'test_alias', 'bid_date': '2023-06-21T10:00:00Z', 'bid_folder_url': 'test_folder_url', 'status': 'in-progress', 'links': {'self': '/bids/99f38b86-ae2c-405c-8508-a897a3521551', 'questions': '/bids/99f38b86-ae2c-405c-8508-a897a3521551/questions'}, 'was_successful': 'test_success', 'success': [{'phase': 1, 'has_score': True, 'score': 80, 'out_of': 100}], 'failed': {'phase': 2, 'has_score': False}, 'feedback': {'description': 'test_description', 'url': 'test_url'}, 'last_updated': ''}
